+++ src/protocol/dap.h
@@ -6,7 +6,6 @@
 #pragma once
 
 #include "types/types.h"
-#include "utils/string_view.h"
 #include <json/json.hpp>
 #include <atomic>
 #include <condition_variable>
@@ -47,7 +46,7 @@ class DAP
     void EmitContinuedEvent(ThreadId threadId);
     void EmitThreadEvent(const ThreadEvent &event);
     void EmitModuleEvent(const ModuleEvent &event);
-    void EmitOutputEvent(OutputCategory category, const Utility::string_view &output, DWORD threadId = 0);
+    void EmitOutputEvent(const OutputEvent &event);
     void EmitBreakpointEvent(const BreakpointEvent &event);
     void CommandLoop();