+++ test-suite/TestStdIO/Program.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.IO;
+using System.Diagnostics;
 
 using DbgTest;
 using DbgTest.DAP;
@@ -21,6 +22,7 @@ class Program
                 Context.AddBreakpoint(@"__FILE__:__LINE__", "bp3");
                 Context.AddBreakpoint(@"__FILE__:__LINE__", "bp4");
                 Context.AddBreakpoint(@"__FILE__:__LINE__", "bp5");
+                Context.AddBreakpoint(@"__FILE__:__LINE__", "bp6");
                 Context.SetBreakpoints(@"__FILE__:__LINE__");
                 Context.PrepareEnd(@"__FILE__:__LINE__");
                 Context.WasEntryPointHit(@"__FILE__:__LINE__");
@@ -45,6 +47,10 @@ class Program
         Console.Error.WriteLine("test \001\002 forbidden \036\037 chars");
 
         i++;                                                      Label.Breakpoint("bp5");
+
+        Debugger.Log(0, "Info", "Application started.");
+
+        i++;                                                      Label.Breakpoint("bp6");
 
         Label.Checkpoint("testio", "finish",
             (Object context) =>
@@ -76,8 +82,11 @@ class Program
                 Context.Continue(@"__FILE__:__LINE__");
 
                 Context.WasBreakpointHit(@"__FILE__:__LINE__", "bp5");
-                // FIXME
                 Context.WasOutputEvent("stderr", "test \u000001\u000002 forbidden \u000036\u000037 chars" + endLine, @"__FILE__:__LINE__");
+                Context.Continue(@"__FILE__:__LINE__");
+
+                Context.WasBreakpointHit(@"__FILE__:__LINE__", "bp6");
+                Context.WasOutputEvent("stdout", "Application started.", @"__FILE__:__LINE__");
                 Context.Continue(@"__FILE__:__LINE__");
             });